+++ src/app/pages/ErrorDetails.jsx
@@ -39,10 +39,16 @@ export default class ErrorDetails extends React.Component {
       </div>;
     }
 
+    const importerror = ed && ed.scriptUrl && ed.scriptUrl.indexOf('/import') != -1;
+
     return <div className='error'>
       <h1>Jameson, we have a problem..</h1>
       <h1><small>{error.message}</small></h1>
+      <br/>
+        {importerror ? <div>If you are attempting to import a ship from EDDI or EDMC and are seeing a 'Z_BUF_ERROR' it means that the URL has not been provided correctly.  This is a common problem when using Microsoft Internet Explorer or Microsoft Edge, and you should use another browser instead.</div> : null }
+      <br/>
       <div>Please note that this site uses Google Analytics to track performance and usage.  If you are blocking cookies, for example using Ghostery, please disable blocking for this site and try again.</div>
+      <br/>
       {content}
     </div>;
   }